서로 테스트

+++ src/main/java/org/zerock/domain/ReplyVo.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class ReplyVo {
 	private int REPLY_NO;
-	private int BOARD_NO;
+	private Long BOARD_NO;
 	private int MEM_NO;
 	private String REPLY_CONTENT;
 	private Date REPLY_DATE;
